fix(cli-host): name what the pricing line counts, not models
`pricing_finding` formatted the vendored snapshot's rule count as
"633 models". `nika catalog`, one command away on a first session,
says "69 models". Two inventories under one word, an order of
magnitude apart, and nothing on either surface told the reader they
were different questions (#1179).

The counts were never wrong; the facets were unnamed — the same
RAMS-12 / A-06 disease `machine_truth.rs` exists to cure, one
surface further out. The line now reads "633 price rules · 10
providers priced": several patterns can price one model, and a
pattern can price models this catalog never lists, so neither count
contains the other.

The test asserts the whole class, not the instance: the pricing
detail must not speak the word `models` at all. Mutation-proven —
restoring the old format string turns it red.

`doctor/tests.rs` sat one line under the 1500-LOC wall, so the
pricing pins descend into a `pricing_tests` sibling module rather
than buy an exemption (the `lsp_transport.rs` precedent).
