+++ changelog.d/1179.changed.md
@@ -1,0 +1,7 @@
+- **`nika doctor` stops calling price rows models (#1179).** The pricing
+  line formatted the vendored snapshot's rule count as « 633 models »,
+  one command away from `nika catalog`'s « 69 models » — two inventories
+  under one word, an order of magnitude apart, both on a first session.
+  It now reads `633 price rules · 10 providers priced`. Same numbers,
+  each naming the facet it counts: several patterns can price one model,
+  and a pattern can price models this catalog never lists.
